+++ app/spdk_vrdma/README.md
@@ -29,9 +29,10 @@
  *   (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  *   OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-===========================
+
 How to use SPDK test on ARM 
 ===========================
+
 Basic BF2 version
 ==================
 DOCA_v1.1_BlueField_OS_CentOS_7.6-5.4.0-1013.11.gc93744f-bluefield-5.4-0.6.5.0-3.7.0.11797-1-aarch64
@@ -53,29 +54,19 @@ How to build
 
 How to run
 ==================
-# Since we hardcode use virtio-net before FW ready, we need configure net in NVCONFIG
-# mlxconfig -d /dev/mst/mt41686_pciconf0 -y s VIRTIO_NET_EMULATION_ENABLE=1 \
-#                      PCI_SWITCH_EMULATION_ENABLE=0 \
-#                      NVME_EMULATION_ENABLE=0 \
-#                      VIRTIO_NET_EMULATION_NUM_PF=1 \
-#                      VIRTIO_NET_EMULATION_NUM_VF=4 \
-#                      VIRTIO_NET_EMULATION_NUM_MSIX=16 \
-#                      NUM_VF_MSIX=64  NUM_OF_VFS=16 \
-#                      PF_BAR2_ENABLE=0 PER_PF_NUM_SF=1 PF_TOTAL_SF=150 PF_SF_BAR_SIZE=10 \
-#                      SRIOV_EN=1
-# Note: need FW reset to let NVCONFIG work.
+mlxconfig -d /dev/mst/mt41686_pciconf0 r
+#Note: need FW reset to let NVCONFIG work.
 
 export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/usr/local/snap/lib #will be deleted this step in release version
 echo 4096 > /sys/devices/system/node/node0/hugepages/hugepages-2048kB/nr_hugepages
-# systemctl stop virtio-net-controller  #will be deleted this step after FW ready
 
 <spdk_vrdma_view>./app/spdk_vrdma/spdk_vrdma
 
 #Note:Must run spdk on ARM before dpdk on host to make sure device reset successfully.
 
-=================================
 How to use DPDK test on host 
 =================================
+
 How to get code
 ==============================
 git clone git@github.com:LiZhang-2020/dpdk.org.git  dpdk_vrdma_view
@@ -92,7 +83,7 @@ How to run testpmd
 echo 4096 > /sys/devices/system/node/node0/hugepages/hugepages-2048kB/nr_hugepages
 
 #2. Find vrdma device pci, such as 5e:00.3
-lspci |grep Tencent
+lspci |grep 4ace
 5e:00.3 Ethernet controller: Tencent Technology (Shenzhen) Company Limited Device 4ace
 
 #3. Bond vrmda device to dpdk manage it, such as 0000:5e:00.3
@@ -103,10 +94,10 @@ sudo ./build/app/dpdk-testpmd --log-level=.,8 -a 5e:00.3 --iova-mode=pa  -- -i -
 
 How to run testpmd command
 ==============================
-create vrdma adminq 5e:00.3 0         #Will get admin-queue by vrdma device 5e:00.3
-create vrdma adminq msg 0 106         #Will create PD index 0 for vrdma device on ARM.
-create vrdma qp 0 4                   #Will 4 qps for test traffic
-#dump vrdma adminq 0           #Will dump admin-queue message
+create vrdma adminq 5e:00.3 0    #Will get admin-queue by vrdma device 5e:00.3
+create vrdma adminq msg 0 106    #Will create PD index 0 for vrdma device on ARM.
+create vrdma qp 0 4              #Will 4 qps for test traffic
+#dump vrdma adminq 0             #Will dump admin-queue message
 #del vrdma adminq 0
 
 How to run RPC command on ARM
@@ -118,7 +109,7 @@ snap-rdma/rpc/snap_rpc.py controller_vrdma_configue -d 0 -e mlx5_0  -v 2 -b 2 -c
 
 How to run testpmd command for qp ready
 ==============================
-# modify vrdma qp <dev_id> <qp_idx> <qp_state> <dest_qpn> <sip> <dip>
+#modify vrdma qp <dev_id> <qp_idx> <qp_state> <dest_qpn> <sip> <dip>
 One example:
 modify vrdma qp 0 2 1 2 0x1234 0x5678  #Set qp to init
 modify vrdma qp 0 2 2 2 0x1234 0x5678  #Set qp to RTR
