+++ lib/vrdma/vrdma_controller.c
@@ -372,9 +372,3 @@ void vrdma_ctrl_destroy(void *arg, void (*done_cb)(void *arg),
     vrdma_ctrl_free(ctrl);
     SPDK_NOTICELOG("lizh vrdma_ctrl_destroy...done");
 }
-
-uint8_t vrdma_get_qp_status(uint32_t qp_handle)
-{
-	//TODO: get qp based on handle
-	return SW_VIRTQ_SUSPENDED;
-}
